// interfaces/attractionDetails.ts
// Generated by https://quicktype.io

export interface AttractionDetails {
  name: string;
  type: string;
  id: string;
  test: boolean;
  url: string;
  locale: string;
  externalLinks: ExternalLinks;
  aliases: string[];
  images: Image[];
  classifications: Classification[];
  upcomingEvents: UpcomingEvents;
  _links: Links;
}

export interface Links {
  self: Self;
}

export interface Self {
  href: string;
}

export interface Classification {
  primary: boolean;
  segment: Genre;
  genre: Genre;
  subGenre: Genre;
  type: Genre;
  subType: Genre;
  family: boolean;
}

export interface Genre {
  id: string;
  name: string;
}

export interface ExternalLinks {
  youtube: Facebook[];
  twitter: Facebook[];
  itunes: Facebook[];
  facebook: Facebook[];
  spotify: Facebook[];
  musicbrainz: Musicbrainz[];
  instagram: Facebook[];
  homepage: Facebook[];
}

export interface Facebook {
  url: string;
}

export interface Musicbrainz {
  id: string;
}

export interface Image {
  ratio: Ratio;
  url: string;
  width: number;
  height: number;
  fallback: boolean;
}

export enum Ratio {
  The16_9 = '16_9',
  The3_2 = '3_2',
  The4_3 = '4_3',
}

export interface UpcomingEvents {
  _total: number;
  _filtered: number;
}